add register and login

+++ routes/web.php
@@ -18,23 +18,37 @@ use App\Models\Listing;
 Route::get('/', [ListingController::class,'index']);
 
 //Show Create Form
-Route::get('/listings/create', [ListingController::class,'create']);
+Route::get('/listings/create', [ListingController::class,'create'])->middleware('auth');
 
 //Store Listing Data
-Route::post('/listings', [ListingController::class,'store']);
+Route::post('/listings', [ListingController::class,'store'])->middleware('auth');
 
 //Show edit form
-Route::get('/listings/{listing}/edit',[ListingController::class,'edit']);
+Route::get('/listings/{listing}/edit',[ListingController::class,'edit'])->middleware('auth');
 
 //Update listing
-Route::put('listings/{listing}',[ListingController::class,'update']);
+Route::put('listings/{listing}',[ListingController::class,'update'])->middleware('auth');
 
 //Delete listing
-Route::delete('listings/{listing}',[ListingController::class,'destroy']);
+Route::delete('listings/{listing}',[ListingController::class,'destroy'])->middleware('auth');
 
 //Single Listing
 Route::get('/listings/{listing}', [ListingController::class,'show']);
 
+//Show Register/Create Form
+Route::get('/register', [UserController::class,'create'])->middleware('guest');
+
+//Create New User
+Route::post('/users',[UserController::class,'store']);
+
+//Logout
+Route::post('/logout', [UserController::class,'logout'])->middleware('auth');
+
+//Show Login Form
+Route::get('/login', [UserController::class,'login'])->name('login')->middleware('guest');
+
+//Login
+Route::post('/users/authenticate', [UserController::class,'authenticate']);
 
 // Common Resource Routes:
 // index - Show all listings
